Operator: Split the common has_permission for OperatorApi
Move the has_permission from the existing
OperatorAPIParkingPermission as this method of the permission
will be shared among other Operator endpoints too.
